requirements.txt
reportlab>=4.0.0
Pillow>=10.0.0
pdf2image>=1.16.0
qrcode[pil]>=7.4.0
numpy>=1.24.0
pypdfium2>=4.0.0  # Optional: for CMYK verification
